[REFACTOR] #138: 리팩토링
