+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter } from 'react-router-dom';
 import NavBar from './components/layout/NavBar';
-import { CustomRouter } from './components/router/CustomRouter';
 import PriceStaticBar from './components/priceStaticBar/PriceStaticBar';
+import CustomRouter from './components/router/CustomRouter';
 
 function App() {
   return (
