+++ frontend/src/components/common/banner/Banner.tsx
@@ -41,11 +41,11 @@ const InfoWrapper = styled.div`
 `;
 
 const SubTitle = styled.p`
-  color: ${(props) => props.theme.color.gray900};
+  color: ${({ theme }) => theme.color.gray900};
   ${BodyKrRegular3}
 `;
 
 const Title = styled.p`
-  color: ${(props) => props.theme.color.primaryColor700};
+  color: ${({ theme }) => theme.color.primaryColor700};
   ${HeadingKrBold1}
 `;
