+++ frontend/src/pages/ModelTypePage.tsx
@@ -1,11 +1,13 @@
 import ModelBannerContainer from '../containers/ModelTypePage/ModelBannerContainer';
+import ModelFooterContainer from '../containers/ModelTypePage/ModelFooterContainer';
 import ModelSelectContainer from '../containers/ModelTypePage/ModelSelectContainer';
 
-export function ModelTypePage() {
+export default function ModelTypePage() {
   return (
     <>
       <ModelBannerContainer />
       <ModelSelectContainer />
+      <ModelFooterContainer />
     </>
   );
 }
