+++ frontend/src/styles/reset.css
@@ -124,13 +124,21 @@ table {
 }
 
 button {
+  padding: 0;
+  margin: 0;
   background-color: transparent;
   border: none;
   cursor: pointer;
 }
 
+li {
+  padding: 0;
+  margin: 0;
+}
+
 input {
-  margin: 0px;
+  margin: 0;
+  padding: 0;
 }
 
 * {
