+++ frontend/src/utils/constants.ts
@@ -1,0 +1,4 @@
+export const MAX_PAGE = 3;
+export const NUM_IN_A_PAGE = 4;
+export const MAX_TEXT_CNT = 98;
+Object.freeze({ MAX_PAGE, NUM_IN_A_PAGE, MAX_TEXT_CNT });
